requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.8
transformers>=4.44
joblib>=1.3
python-dotenv>=1.0
plotly>=5.24
python-docx>=1.1.0
pdfminer.six>=20231228
openpyxl>=3.1.0
xlrd>=2.0.0
reportlab>=4.0.0
markdown>=3.5.0
requests>=2.31.0

